moving store to components
