+++ src/components/pages/AllCakes.jsx
@@ -1,5 +1,5 @@
 import { Row, Card, Container, Col, Button } from "react-bootstrap";
-import { useCakeStore } from "../../store/cakeStore";
+import { useCakeStore } from "../store/cakeStore";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import DeleteConfirmation from "../shared/DeleteConfirmation";
